requirements.txt
fastapi
uvicorn
httpx
pytest
python-multipart
Pillow